+++ resources/js/registration.js
@@ -4,9 +4,9 @@ import '@/router';
 import '@/datatables';
 
 import '@/components/editable';
-import '@/components/relation';
 import '@/components/password';
 
 import '@/plugins/multilingual';
 import '@/plugins/slugify';
 import '@/common';
+import '@/sidebar';