+++ resources/views/datatables/filters/select.blade.php
@@ -1,20 +1,15 @@
-<select
-        @class([
-            'form-control',
-            'select2'.($filter->isAjax()?'-ajax':'')
-        ])
-        id="{{ $filter->getName() }}"
-        data-width="100%"
-        data-action="dataTableFilterSelect"
-        data-allow-clear="true"
-        data-placeholder="{{ $field->get('label') }}"
-        data-minimum-results-for-search="10"
-        @if($filter->isMultiple()) multiple @endif
-        {!! $filter->convertAttributesToHtml() !!}
-    >
-    @foreach ($filter->getOptions() as $key => $option)
-    <option value="{{ $key }}"
-        @selected(session('datatable.'.$dataType->getSlug().'.'.$field->get('name')) === $key)
-    >{{ $option }}</option>
+{{-- Select-фильтр колонки. Привязан к данным таблицы (datatable.blade.php). --}}
+<el-select
+    v-model="filters['{{ $filter->getName() }}']"
+    size="small"
+    clearable
+    filterable
+    @if($filter->isMultiple()) multiple @endif
+    placeholder="{{ $field->get('label') }}"
+    style="width: 100%"
+    @change="onFilter"
+>
+    @foreach($filter->getOptions() as $key => $label)
+        <el-option value="{{ $key }}" label="{{ $label }}"></el-option>
     @endforeach
-</select>
+</el-select>